user must be logged in to save or delete games, server confirms with auth0 token

+++ server/routes/savedGames.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express'
 import * as db from '../db/savedGames.ts'
 import { GameData } from '../../models/savedGame.ts'
+import checkJwt, { JwtRequest } from '../auth0.ts'
 
 const router = Router()
 
@@ -30,9 +31,18 @@ router.get('/game/:id', async (req, res) => {
   }
 })
 
-router.post('/', async (req, res) => {
+router.post('/', checkJwt, async (req: JwtRequest, res) => {
+  const savedGame: GameData = req.body
+  const auth0Id = req.auth?.sub
+  if (!savedGame) {
+    console.error('No game data provided to save')
+    return res.status(400).send('Bad request')
+  }
+  if (!auth0Id) {
+    console.error('No auth0Id')
+    return res.status(401).send('Unauthorized')
+  }
   try {
-    const savedGame: GameData = req.body
     const game = await db.addNewSavedGame(savedGame)
     game[0] ? res.sendStatus(200) : res.sendStatus(500)
   } catch (error) {
@@ -43,10 +53,18 @@ router.post('/', async (req, res) => {
   }
 })
 
-router.delete('/', async (req, res) => {
+router.delete('/', checkJwt, async (req: JwtRequest, res) => {
+  const { saveId } = req.body
+  const auth0Id = req.auth?.sub
+  if (!saveId) {
+    console.error('No game save id provided')
+    return res.status(400).send('Bad request')
+  }
+  if (!auth0Id) {
+    console.error('No auth0Id')
+    return res.status(401).send('Unauthorized')
+  }
   try {
-    const { saveId } = req.body
-    console.log(saveId)
     const isDeleted = await db.deleteSavedGame(saveId)
     isDeleted ? res.sendStatus(200) : res.sendStatus(500)
   } catch (error) {
